requirements.txt
numpy
opencv-python
python-dotenv
aiohttp
beautifulsoup4
ydb
fake-useragent
ydb[yc]